deprecation ADAL init
